+++ rust/src/examples.rs
@@ -25,81 +25,17 @@ pub fn example_circuit_inputs(x: u32, y: u32, zz: u32) -> CircuitOwned {
 
 
 pub fn write_example_constraints<W: io::Write>(mut writer: W) -> io::Result<()> {
-    let constraints: Vec<((Vec<u64>, Vec<u8>), (Vec<u64>, Vec<u8>), (Vec<u64>, Vec<u8>))> = vec![
+    let constraints_vec: Vec<((Vec<u64>, Vec<u8>), (Vec<u64>, Vec<u8>), (Vec<u64>, Vec<u8>))> = vec![
         // (A ids values)  *  (B ids values)  =  (C ids values)
         ((vec![1], vec![1]), (vec![1], vec![1]), (vec![4], vec![1])),       // x * x = xx
         ((vec![2], vec![1]), (vec![2], vec![1]), (vec![5], vec![1])),       // y * y = yy
         ((vec![0], vec![1]), (vec![4, 5], vec![1, 1]), (vec![3], vec![1])), // 1 * (xx + yy) = z
     ];
 
-    let constraints_owned = from_vec(constraints);
+    let constraints_owned: ConstraintSystemOwned = constraints_vec.into();
 
-    write(&mut writer, constraints_owned)
+    constraints_owned.write(&mut writer)
 }
-
-fn from_vec(constraints: Vec<((Vec<u64>, Vec<u8>), (Vec<u64>, Vec<u8>), (Vec<u64>, Vec<u8>))>) -> ConstraintSystemOwned {
-    // create ConstraintSystemOwned
-    let mut constraints_owned = ConstraintSystemOwned {
-        constraints: vec![]
-    };
-
-    for (lca, lcb, lcc) in constraints {
-        let lca = VariablesOwned {
-            variable_ids: lca.0,
-            values: Some(lca.1),
-        };
-        let lcb = VariablesOwned {
-            variable_ids: lcb.0,
-            values: Some(lcb.1),
-        };
-        let lcc = VariablesOwned {
-            variable_ids: lcc.0,
-            values: Some(lcc.1),
-        };
-
-        constraints_owned.constraints.push(BilinearConstraintOwned {
-            linear_combination_a: lca,
-            linear_combination_b: lcb,
-            linear_combination_c: lcc,
-        });
-    }
-    constraints_owned
-}
-
-fn write<W: io::Write>(writer: &mut W, constraints_owned: ConstraintSystemOwned) -> io::Result<()> {
-// pack as flat buffer
-    let mut builder = &mut FlatBufferBuilder::new();
-    let mut constraints_built = vec![];
-
-    for bilinear_constraints in constraints_owned.constraints {
-        let lca = bilinear_constraints.linear_combination_a.build(builder);
-        let lcb = bilinear_constraints.linear_combination_b.build(builder);
-        let lcc = bilinear_constraints.linear_combination_c.build(builder);
-
-        constraints_built.push(BilinearConstraint::create(builder, &BilinearConstraintArgs {
-            linear_combination_a: Some(lca),
-            linear_combination_b: Some(lcb),
-            linear_combination_c: Some(lcc),
-        }));
-    }
-
-    let constraints_built = builder.create_vector(&constraints_built);
-    let r1cs = ConstraintSystem::create(&mut builder, &ConstraintSystemArgs {
-        constraints: Some(constraints_built),
-        info: None,
-    });
-
-    let message = Root::create(&mut builder, &RootArgs {
-        message_type: Message::ConstraintSystem,
-        message: Some(r1cs.as_union_value()),
-    });
-    builder.finish_size_prefixed(message, None);
-
-    writer.write_all(builder.finished_data())
-}
-
-
-
 
 pub fn write_example_witness<W: io::Write>(writer: W) -> io::Result<()> {
     write_example_witness_inputs(writer, 3, 4)
